change app name and logo

+++ app/Http/Controllers/Admin/ApplicationController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Admin;
+use App\Application;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ChangeAdminPassword;
@@ -11,11 +12,40 @@ use App\User;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
-class SettingController extends Controller
+class ApplicationController extends Controller
 {
     public function index()
     {
         return view('admin.settings', ['user' => Auth::user()]);
+    }
+
+    public function updateApplication(Request $request, Application $application)
+    {
+        if ($request->has('logo')) {
+
+            // Get image extension
+            $logo = $request->logo;
+            $ext = $logo->getClientOriginalExtension();
+
+            // store in public folder
+            $filename = uniqid() . '.' .$ext;
+            $logo->storeAs('public/logos', $filename);
+
+            // delete existing user image
+            if (isset($application->logo) && $application->logo != 'default.jpg') {
+                Storage::delete("/public/logos/" . $application->logo);
+            }
+
+            // assign to request
+            $request->logo = $filename;
+        }
+
+        // update
+        $application->name = $request->name;
+        $application->logo = $request->logo ?: $application->logo;;
+        $application->save();
+
+        return back()->with('success', 'Successfully Updated Settings');
     }
 
     public function updateProfile(UpdateAdminProfile $request, Admin $user)
@@ -31,7 +61,7 @@ class SettingController extends Controller
             $avatar->storeAs('public/avatars', $filename);
 
             // delete existing user image
-            if (isset($user->avatar)) {
+            if (isset($user->avatar) && $user->avatar != 'default.jpg') {
                 Storage::delete("/public/avatars/" . $user->avatar);
             }
 
@@ -42,7 +72,7 @@ class SettingController extends Controller
         // update
         $user->name = $request->name;
         $user->email = $request->email;
-        $user->avatar = $request->avatar;
+        $user->avatar = $request->avatar ?: $user->avatar;
         $user->save();
 
         return back()->with('success', 'Successfully Updated Profile');
